Add 'friend' as 2nd author; appending a link notifies the feed
